最终使用子组建方案
